requirements.txt
alembic==1.18.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
aiosqlite==0.21.0
asn1crypto==1.5.1
click==8.3.1
colorama==0.4.6
fastapi==0.128.0
greenlet==3.3.0
h11==0.16.0
httptools==0.7.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
pg8000==1.31.5
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.1.0
scramp==1.4.8
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
watchfiles==1.1.1
websockets==16.0
passlib[bcrypt]
requests
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
pyotp
fpdf2
qrcode[pil]
pydantic[email]>=1.10.0
psycopg[binary]
prometheus-fastapi-instrumentator==7.1.0
cryptography>=42.0.0
hvac==2.3.0
celery==5.4.0
numpy==2.1.3
scikit-learn==1.5.2
xgboost==2.1.1
joblib==1.4.2
opentelemetry-sdk==1.27.0
opentelemetry-api==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-requests==0.48b0
opentelemetry-exporter-otlp==1.27.0
python-json-logger==2.0.7
httpx==0.27.2
aiosmtplib==3.0.1
pytest-cov==5.0.0
boto3==1.34.162
pytest-asyncio==0.24.0
